feat(dynamics): (semi-)flows, omega limits (#4843)
This code has gone through a couple of iterations since it was first written in summer, when the ambition was 'Morse decompositions in Lean' rather than 'mildly generalise some results from a first course in differential equations'. Nevertheless there's much in here I'm not confident about & would appreciate help with.

Co-authored-by: Yury G. Kudryashov <urkud@urkud.name>